Added signal for which Rule derives atom
Upon clicking an atom, the rule that derives it is highlighted in the
rule header, aside from showing the arrows to reason atoms.

To find the rule that derives an atom from a component with multiple
rules, the justifier program was modified:

````
h(I, J, d, r) :- d, r.
```
where I is the index of the component, J is the index of the rule within
the component, d is the atom, and r are the reason atoms.

The node dataclass stores a dictionary mapping the atoms in the node
to the hash of the rule it derives from. This attribute of the node is
queried through the API when an atom is clicked.

The React App then stores the highlighted rules in the context
HighlightedSymbol in the shape
```
HighlightedRule { rule_hash: string, color: string, source_id: string }
```

Contributes: potassco/viasp#22