Justifier program rule hash optimization
For the justifier program, we have a rule hash that is used to identify
the rule that justifies an atom. This hash is computed by hashing the
string of the rule. This ensures that even for multiple dependants
of the same rule, the hash is always the same.

Contributes: potassco/viasp#22