Add non-ground search
The backend assumes the input query is a non-ground atom, if the string
search does not return any results.

The ground instances of the input atom are found by grounding
the program:
`
model(hc(1,2),'id1').model(hc(2,3), 'id2'). %...all symbols of graph
result({query},SYMBOLUUID):-model({query},SYMBOLUUID).
`
If a syntax error breaks grounding, the input is assumed to be
incomplete.

The frontend shows a loading dots animation to signify that the search
query is assumed to be incomplete.

Contributes: potassco/viasp#85