Partial fix for include statement.
Does not work in other directory yet.

To recognize #include statements, the viasp pythonAPI
parses the input file and checks the ast locations.
Each separate filename and its content is registered to the server using
the load call. The server stores multiple files in the
Encodings Database with their filenames and contents.
The encoding_id is the session_id for the encoding.

When creating the dependency graph, each input file is
parsed separately. A encoding_map is used to fetch the input
string. This map keys the filename to the file contents.
Transformations of the dependency graph are built using the input
string fetched from the encoding_map.

To ensure the #include statements are skipped when analyzing the
program, the visit function was modified. The original input program
is called <string>, whereas the included files are called by their
filename.