Load files referenced with `#include` (#93)
To recognize #include statements, the viasp pythonAPI
parses the input file and checks the values of the ast's location variable.
Each separate filename represents either the input file, or a file importet
by a `include` statement in an input file.

Each filename and its content is registered to the server using
the load call. The server stores all encoding files in the
`Encodings` Table. The encoding_id is the session_id for the encoding.

In the Program Analysis step, each input file is
parsed separately. In order to store the rules' input string next to the AST, the 
input program is passed as a `kwarg` to the `ProgramAnalyzer`'s visitor.
The AST, the input string and a identifying hash are stored in the `RuleContainer`.

The dependency graph is built using the `RuleContainer`. Loops or cycles are slashed
by merging `RuleContainer`.

To ensure the #include statements are skipped when the analyzer parses the
program, the visit function was modified. Every direct input program
(filename is always `<string>`) is parsed. Indirect input programs are called by their
filename and are skipped.

Fixes: potassco/viasp#92