Sort row of constraints last in primary Sort
On initial graph generation, components (rows of the graph) that contain
rules of only integrity constraints are now moved to the end of the
graph.

This is done to avoid unnecessary clutter on the graph, as
integrity constraints are only shown for completeness and do not play
a role in the justification process.

However, through manual reordering using drag & drop in the client, the
constraints can be moved to the desired position in the graph.

To weaken or disable this feature, edit the function
``move_constraints_only_transformations_to_end`` in the file
``../backend/src/viasp/asp/utils.py``.