Serve the frontend with flask, manage shutdown
To differentiate between different sessions (e.g. multiple terminals
call viasp for different encodings), each session is assigned a unique
token. The link brings the user from the terminal to the browser with
the token as a query parameter. The token is then stored in a recoil
atom and sent to the backend server with each API call.

When viasp is interrupted, an API call asks the backend server
if there are other active viasp sessions. If not, the backend server
and frontend server are shut down. Otherwise, no shutdown happens, but
the current session is deregistered from the backend server.

The startup process is sped up by starting the server before the
generation of the viasp graph.

Resolves: potassco/viasp#67